Move ref counted stuff to krEngine/ownership
