+++ code/krEngine/ownership.h
@@ -3,3 +3,6 @@
 #include <krEngine/ownership/ownershipData.h>
 #include <krEngine/ownership/owned.h>
 #include <krEngine/ownership/borrowed.h>
+
+#include <krEngine/ownership/refCounted.h>
+#include <krEngine/ownership/refCountedPtr.h>
