+++ code/krEngine/ownership/refCounted.h
