+++ code/krEngine/rendering/implementation/window.cpp
@@ -1,6 +1,6 @@
 #include <krEngine/pch.h>
 #include <krEngine/rendering/window.h>
-#include <krEngine/referenceCounting.h>
+#include <krEngine/ownership.h>
 
 #include "windowImpl.h"
